everything...

+++ src/commands/game/card/UpgradeCard.ts
@@ -11,8 +11,12 @@ import { GameUserCard } from "../../../structures/game/UserCard";
 import { Dust } from "../../../structures/game/Dust";
 import { items } from "../../../assets/Items";
 import { getDescriptions } from "../../../lib/utility/text/TextUtils";
-import { checkPermission, isDeveloper } from "../../../lib/ZephyrUtils";
+import { checkPermission } from "../../../lib/ZephyrUtils";
 import { Zephyr } from "../../../structures/client/Zephyr";
+import { QuestGetter } from "../../../lib/database/sql/game/quest/QuestGetter";
+import { QuestSetter } from "../../../lib/database/sql/game/quest/QuestSetter";
+import { QuestObjective } from "../../../structures/game/quest/QuestObjective";
+import { QuestProgression } from "../../../structures/game/quest/QuestProgression";
 
 export default class UpgradeCard extends BaseCommand {
   id = `sunshine`;
@@ -26,9 +30,6 @@ export default class UpgradeCard extends BaseCommand {
     profile: GameProfile,
     options: string[]
   ): Promise<void> {
-    if (!Zephyr.flags.upgrades && !isDeveloper(msg.author))
-      throw new ZephyrError.UpgradeFlagDisabledError();
-
     const reference = options[0];
     let card: GameUserCard;
     if (!reference) {
@@ -45,9 +46,11 @@ export default class UpgradeCard extends BaseCommand {
     if (card.wear === 5) throw new ZephyrError.CardBestConditionError(card);
 
     const dustTier = (card.wear + 1) as Dust;
-    const dustCost = 5;
+    const dustCost = Zephyr.config.testing ? 0 : 5;
     const successChance = [80, 70, 60, 50, 40][card.wear];
-    const bitCost = [75, 150, 250, 400, 600][card.wear];
+    const bitCost = Zephyr.config.testing
+      ? 0
+      : [75, 150, 250, 400, 600][card.wear];
 
     if (bitCost > profile.bits)
       throw new ZephyrError.NotEnoughBitsError(bitCost);
@@ -55,29 +58,45 @@ export default class UpgradeCard extends BaseCommand {
     const dustItem = items.filter((i) => i.names[0].includes("Dust"))[
       card.wear
     ];
-    const dustUserItem = await ProfileService.getItem(
-      profile,
-      dustItem.id,
-      dustItem.names[0]
-    );
 
-    if (dustCost > dustUserItem.quantity)
-      throw new ZephyrError.NotEnoughDustError(
-        dustUserItem.quantity,
-        dustCost,
-        dustTier
+    if (dustCost > 0) {
+      const dustUserItem = await ProfileService.getItem(
+        profile,
+        dustItem.id,
+        dustItem.names[0]
       );
+
+      if (dustCost > dustUserItem.quantity)
+        throw new ZephyrError.NotEnoughDustError(
+          dustUserItem.quantity,
+          dustCost,
+          dustTier
+        );
+    }
+
+    let cost = `This upgrade will cost...\n`;
+
+    if (dustCost === 0 && bitCost === 0) {
+      cost = `This upgrade is free.`;
+    } else {
+      if (dustCost > 0) {
+        cost += `\n— **${dustCost.toLocaleString()}x** \`${
+          dustItem.names[0]
+        }\``;
+      }
+      if (bitCost > 0) {
+        cost += `\n— **${
+          Zephyr.config.discord.emoji.bits
+        } ${bitCost.toLocaleString()}**`;
+      }
+    }
 
     const tags = await ProfileService.getTags(profile);
     const embed = new MessageEmbed(`Upgrade`, msg.author)
       .setDescription(
         `Are you sure you want to upgrade this card?` +
           `\n${await getDescriptions([card], tags)}` +
-          `\n\nThis will cost...` +
-          `\n— **${dustCost.toLocaleString()}x** \`${dustItem.names[0]}\`` +
-          `\n— **${
-            Zephyr.config.discord.emoji.bits
-          } ${bitCost.toLocaleString()}**`
+          `\n\n${cost}`
       )
       .setFooter(`🎲 Chance of success: ${successChance}%`);
 
@@ -128,10 +147,14 @@ export default class UpgradeCard extends BaseCommand {
     if (refetchCard.wear !== card.wear)
       throw new ZephyrError.UpgradeFailedError();
 
-    await ProfileService.removeItems(profile, [
-      { item: dustItem, count: dustCost },
-    ]);
-    await ProfileService.removeBitsFromProfile(profile, bitCost);
+    if (dustCost > 0) {
+      await ProfileService.removeItems(profile, [
+        { item: dustItem, count: dustCost },
+      ]);
+    }
+    if (bitCost > 0) {
+      await ProfileService.removeBitsFromProfile(profile, bitCost);
+    }
 
     const chance = new Chance();
     const success = chance.bool({ likelihood: successChance });
@@ -141,6 +164,21 @@ export default class UpgradeCard extends BaseCommand {
       await confirmation.edit({
         embed: embed.setFooter(`🎉 The upgrade succeeded!`),
       });
+
+      if (refetchCard.wear + 1 === 5) {
+        const progressableQuests = await QuestGetter.checkAvailableQuestsForProgress(
+          profile,
+          QuestObjective.UPGRADE_TO_MINT
+        );
+
+        if (progressableQuests.length > 0) {
+          const progressions = progressableQuests.map((q) => {
+            return { ...q, increment: 1 } as QuestProgression;
+          });
+
+          await QuestSetter.progressQuests(progressions, profile);
+        }
+      }
     } else {
       await confirmation.edit({
         embed: embed.setFooter(`😕 The upgrade failed.`),
